backend/upload_function/requirements.txt
azure-functions
azure-storage-blob
azure-servicebus
